+++ src/main/java/com/transloadit/sdk/AssemblyListener.java
@@ -3,6 +3,7 @@ package com.transloadit.sdk;
 import com.transloadit.sdk.response.AssemblyResponse;
 import org.json.JSONObject;
 
+
 /**
  * Interface for a Listener, which tracks the progress in Assembly execution.
  */